# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buildspace API",
    "version": "0.1.0",
    "description": "Separate standalone backend contract. Full request schemas live alongside Hono routes in src/modules."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/v1/spaces/{slug}/chat": {
      "get": {
        "summary": "Space chat history",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/spaces/{slug}/chat/messages": {
      "post": {
        "summary": "Send Space chat message",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/dm/requests": {
      "post": {
        "summary": "Send direct-message request",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "get": {
        "summary": "Message requests",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/dm/requests/{requestId}": {
      "patch": {
        "summary": "Accept or decline message request",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/dm/conversations": {
      "get": {
        "summary": "Direct conversations",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/conversations/{conversationId}/messages": {
      "get": {
        "summary": "Direct message history",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Send direct message",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/messages/{messageId}/read": {
      "post": {
        "summary": "Mark message read",
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/spaces/{slug}/files/upload-url": {
      "post": {
        "summary": "Issue signed upload URL",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/files/{fileId}/download-url": {
      "get": {
        "summary": "Issue signed download URL",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "buildspace_session"
      }
    }
  }
}